Moved all relevant code into the Wiimote plugin.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@518 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/ExtendedTrace.cpp
@@ -430,7 +430,7 @@ void etfprintf(FILE *file, const char *format, ...) {
 }
 
 void etfprint(FILE *file, const std::string &text) {
-	int len = text.length();
+	size_t len = text.length();
 	fwrite(text.data(), 1, len, file);
 }
